+++ src/jcpsantiago/arqivist/api/slack/router.clj
@@ -1,14 +1,15 @@
 (ns jcpsantiago.arqivist.api.slack.router
   "Reitit routes for interacting with Slack."
   (:require
-   [jcpsantiago.arqivist.api.slack.handler :as handler]))
+   [jcpsantiago.arqivist.api.slack.handlers :as handlers]
+   [jcpsantiago.arqivist.api.slack.specs :as specs]))
 
 (defn routes
   "Routes invoked by Slack:
    * shortcut — triggered after the user clicks the message shortcut button
    * slash    — triggered after the user uses the `/arqive` slash command
    * redirect — called as part of the OAuth process at the end of installation"
-  [_]
+  [system]
   ["/slack"
    {:swagger {:tags ["Slack"]}}
    ["/shortcut"
@@ -21,7 +22,7 @@
       ;; TODO: add the rest of the specs, not working yet
       ;; :parameters {:body :jcpsantiago.arqivist.api.slack.spec/shortcut-body}
       :responses {200 {:body string?}}
-      :handler handler/message-shortcut}}]
+      :handler handlers/message-shortcut}}]
 
    ;; TODO: Add example from https://api.slack.com/interactivity/slash-commands#app_command_handling
    ["/slash"
@@ -34,7 +35,7 @@
       ;; TODO: add the rest of the specs, not working yet
       ;; :parameters {:form :jcpsantiago.arqivist.api.slack.spec/slash-form-params}
       :responses {200 {:body string?}}
-      :handler handler/slash-command}}]
+      :handler handlers/slash-command}}]
 
    ["/redirect"
     {:swagger {:externalDocs
@@ -43,7 +44,7 @@
      :get
      {:summary "OAuth2 redirect target"
       :description "This endpoint receives the data about the workspace, after a user successfully added the app to their account."
-      ;; TODO: add the rest of the specs, not working yet
-      ;; :parameters {:body :jcpsantiago.arqivist.api.slack.spec/redirect-body}
-      :responses {200 {:body string?}}
-      :handler {:status 200 :body "HELLO"}}}]])
+      :parameters {:query ::specs/oauth-redirect}
+      :responses {200 {:body string?}
+                  500 {:body string?}}
+      :handler (handlers/oauth-redirect system)}}]])