+++ src/jcpsantiago/arqivist/api/slack/specs.clj
@@ -1,20 +1,68 @@
 (ns jcpsantiago.arqivist.api.slack.specs
-  "Specs for Slack data, including incoming requests, db
-  representations and internal maps."
-  (:require [clojure.spec.alpha :as spec]))
+  "
+  Specs for Slack data, including incoming requests,
+  db representations and internal maps.
+  "
+  (:require
+   [clojure.spec.alpha :as spec]
+   [clojure.string :refer [blank?]]))
 
+;; Extra predicates
+(defn non-blank-string?
+  "
+  Predicate for non blank strings.
+  "
+  [x]
+  (and (string? x) (not (blank? x))))
+
+(spec/def ::access_token non-blank-string?)
+(spec/def ::scope non-blank-string?)
+(spec/def ::bot_user_id non-blank-string?)
+(spec/def ::app_id string?)
+(spec/def ::id string?)
+(spec/def ::name string?)
+(spec/def ::token_type #{"bot" "user"})
+
+;; Error response
 (spec/def ::ok boolean?)
 (spec/def ::error string?)
-(spec/def ::api-error
+(spec/def ::error-response
   (spec/keys
    :req-un [::ok ::error]))
 
+(spec/def ::team
+  (spec/keys
+   :req-un [::id]
+   :opt-un [::name]))
+
+(spec/def ::authed_user
+  (spec/keys
+   :req-un [::id]
+   :opt-un [::scope ::access_token ::token_type]))
+
+;; OAuth redirect -----------------------------------------------
+(spec/def ::code non-blank-string?)
+(spec/def ::state non-blank-string?)
+
+;; Initial request received from Slack once user allows the requested scopes
+;; code is exchanged for a request token (see spec below)
+;; official docs in https://api.slack.com/authentication/oauth-v2
+(spec/def ::oauth-redirect
+  (spec/keys
+   :req-un [::code ::state]))
+
+;; Access token request
+(spec/def ::oauth-access
+  (spec/or
+   :good-response (spec/keys
+                   :req-un [::access_token ::scope ::bot_user_id ::app_id ::team ::authed_user])
+   :error-response ::error-response))
+
+;; App uninstallation
 (spec/def ::apps-uninstall
   (spec/or
    :ok-response (spec/keys :req-un [::ok])
    :error-response ::api-error))
-
-;; Incoming requests -------------------------------------------------------
 
 ;; Slash commands are sent via POST requests with Content-type application/x-www-form-urlencoded.
 ;; See the docs in https://api.slack.com/interactivity/slash-commands#app_command_handling
@@ -60,67 +108,3 @@
   (spec/keys
    :req-un [::id ::uuid ::app_id ::external_team_id ::team_name ::registering_user
             ::scopes ::access_token ::bot_user_id ::created_at]))
-
-;; API responses -----------------------------------------------------------
-
-(spec/def ::next_cursor string?)
-(spec/def ::response_metadata
-  (spec/keys
-   :req-un [::next_cursor]))
-
-(spec/def ::channel
-  (spec/keys
-   :req-un [::id ::name ::name_normalized ::is_channel ::created ::creator
-            ::is_archived ::is_general ::is_shared ::parent_conversation
-            ::is_ext_shared ::is_pending_ext_shared ::is_org_shared ::is_member
-            ::is_private ::is_mpim ::last_read ::topic ::purpose ::previous_names]
-   :opt-un [::warning]))
-
-(spec/def ::convo-info
-  (spec/or
-   ::api-error
-   (spec/keys
-    :req-un [::ok ::channel])))
-
-(spec/def ::convo-members
-  (spec/or
-   ::api-error
-   (spec/keys
-    :req-un [::ok ::members ::response_metadata])))
-
-(spec/def ::profile
-  (spec/keys
-   :req-un [::title ::phone ::skype ::real_name ::real_name_normalized
-            ::display_name ::display_name_normalized ::status_text
-            ::status_emoji ::status_expiration ::email ::first_name
-            ::last_name]))
-
-(spec/def ::user-profile
-  (spec/or
-   ::api-error
-   (spec/keys
-    :req-un [::ok ::profile])))
-
-(spec/def ::messages
-  (spec/keys
-   :req-un [::type ::user ::text ::ts]
-   :opt-un [::attachments ::subtype ::hidden ::is_starred
-            ::pinned_to ::reactions]))
-
-(spec/def ::convo-history
-  (spec/or
-   ::api-error
-   (spec/keys
-    :req-un [::ok ::messages ::has_more ::pin_count ::response_metadata])))
-
-(spec/def ::convo-join
-  (spec/or
-   ::api-error
-   (spec/keys
-    :req-un [::ok ::channel])))
-
-(spec/def ::convo-replies
-  (spec/or
-   ::api-error
-   (spec/keys
-    :req-un [::ok ::messages])))